# docs/engineering_v0.5/api/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "LAST MILE single-player command API",
    "version": "0.5.0",
    "description": "实现级设计契约，非已实现服务。SINGLE_PLAYER_REFERENCE为待审批参考profile；normalStartAllowed=false时只允许明确标识的design_preview。"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "LaunchBearer": []
    }
  ],
...
  "paths": {
...
    "/bootstrap": {
      "get": {
        "operationId": "getBootstrap",
        "tags": [
          "system"
        ],
        "summary": "授权能力、参考profile和版本",
        "description": "不在正文回传bearer。review profile的normalStartAllowed=false。",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功；缓存幂等回执保留相同HTTP状态和正文。",
            "headers": {
              "X-Request-Id": {
                "description": "请求关联ID；所有响应均返回。",
                "schema": {
                  "type": "string",
                  "minLength": 36,
                  "maxLength": 36,
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../contracts/public.schema.json#/$defs/BootstrapView"
                },
                "example": {
                  "launchId": "11111111-1111-4111-8111-111111111111",
                  "apiVersion": "v1",
                  "contractVersion": "0.5",
                  "capabilities": [
                    "commander.read",
                    "session.create",
                    "session.command",
                    "evaluation.request",
                    "export.request"
                  ],
                  "profiles": [
                    {
                      "profileId": "SINGLE_PLAYER_REFERENCE",
                      "status": "review",
                      "mode": "single_player",
                      "normalStartAllowed": false,
                      "designPreviewAllowed": false,
                      "policyHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "contentVersionId": "11111111-1111-4111-8111-111111111111",
                      "description": "REFERENCE"
                    }
                  ],
                  "maximumRequestBytes": 65536,
                  "sseReconnectMs": 2000
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-authorization": "local-launch-bound-commander",
        "x-idempotency": "safe-read"
      }
    },
...
    "/sessions/{sessionId}/events": {
      "get": {
        "operationId": "streamEvents",
        "tags": [
          "session"
        ],
        "summary": "订阅限定视图的SSE事件",
        "description": "fetch流携带Authorization；不使用URL令牌。Last-Event-ID与after只能二选一，格式epoch:viewSequence。游标绑定session/commander视图；过期410后GET投影恢复。保留本局全部已公开事件，进程跨重启从公开outbox恢复。15秒无数据发送注释keepalive。",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "description": "",
            "schema": {
              "type": "string",
              "minLength": 38,
              "maxLength": 64,
              "pattern": "^[0-9a-f-]{36}:[0-9]+$"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string",
              "minLength": 38,
              "maxLength": 64,
              "pattern": "^[0-9a-f-]{36}:[0-9]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功；缓存幂等回执保留相同HTTP状态和正文。",
            "headers": {
              "X-Request-Id": {
                "description": "请求关联ID；所有响应均返回。",
                "schema": {
                  "type": "string",
                  "minLength": 36,
                  "maxLength": 36,
                  "format": "uuid"
                }
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "minLength": 0,
                  "maxLength": 1048576
                },
                "example": "id: 11111111-1111-4111-8111-111111111111:1\nevent: clock.sample\ndata: {\"eventType\":\"clock.sample\",\"sessionId\":\"11111111-1111-4111-8111-111111111111\",\"runEpoch\":\"11111111-1111-4111-8111-111111111111\",\"viewSequence\":1,\"stateVersion\":0,\"data\":{\"missionTimeMs\":0,\"serverNow\":\"2026-09-16T12:00:00Z\",\"missionDeadlineMs\":600000}}\n\n",
                "x-event-schema": {
                  "$ref": "../contracts/public.schema.json#/$defs/PublicSseEvent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "410": {
            "$ref": "#/components/responses/Error410"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-authorization": "local-launch-bound-commander",
        "x-idempotency": "safe-read"
      }
    },
...
    "/sessions/{sessionId}/evaluations": {
      "post": {
        "operationId": "requestEvaluation",
        "tags": [
          "review"
        ],
        "summary": "基于封存行为包请求独立评价",
        "description": "不接收expectedStateVersion，验证sealedHash。每seal+config复用一个logical job；成功返回缓存，不重新计模型调用。首次与最多一次恢复共享2次上限。评价不能写世界。",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RunEpoch"
          }
        ],
        "responses": {
          "202": {
            "description": "成功；缓存幂等回执保留相同HTTP状态和正文。",
            "headers": {
              "X-Request-Id": {
                "description": "请求关联ID；所有响应均返回。",
                "schema": {
                  "type": "string",
                  "minLength": 36,
                  "maxLength": 36,
                  "format": "uuid"
                }
              },
              "Idempotency-Replayed": {
                "description": "缓存回执为true；新执行为false。",
                "schema": {
                  "type": "boolean"
                }
              },
              "Location": {
                "description": "轮询或读取已创建资源；正文operationLocation/location相同。",
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200,
                  "pattern": "^/api/v1/"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../contracts/public.schema.json#/$defs/EvaluationAccepted"
                },
                "example": {
                  "requestId": "11111111-1111-4111-8111-111111111111",
                  "sealedHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "job": {
                    "jobId": "11111111-1111-4111-8111-111111111111",
                    "agentRole": "evaluator",
                    "status": "queued",
                    "mode": "live_model",
                    "inputVersion": 1,
                    "createdAt": "2026-09-16T12:00:00Z",
                    "attemptCount": 0,
                    "result": null,
                    "error": null
                  },
                  "operationLocation": "/api/v1/sessions/11111111-1111-4111-8111-111111111111/evaluations/11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-authorization": "local-launch-bound-commander",
        "x-idempotency": "required; unique key per launch/session; hash canonical body + method + path + runEpoch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../contracts/public.schema.json#/$defs/EvaluationRequest"
              },
              "example": {
                "sealedHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                "evaluationConfigId": "EVALUATION_REFERENCE_V1"
              }
            }
          }
        }
      }
    },
...
    "/sessions/{sessionId}/exports": {
      "post": {
        "operationId": "createExport",
        "tags": [
          "review"
        ],
        "summary": "构造本地可下载复盘JSON",
        "description": "仅sealed；includePlayerStatements需明确选择。导出只含公开复盘/结果及获准评价，不含密钥、提示词、NPC未上报库存或案卷。不得以下载接口读取任意路径。",
        "parameters": [
          {
            "$ref": "#/components/parameters/sessionId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RunEpoch"
          }
        ],
        "responses": {
          "202": {
            "description": "成功；缓存幂等回执保留相同HTTP状态和正文。",
            "headers": {
              "X-Request-Id": {
                "description": "请求关联ID；所有响应均返回。",
                "schema": {
                  "type": "string",
                  "minLength": 36,
                  "maxLength": 36,
                  "format": "uuid"
                }
              },
              "Idempotency-Replayed": {
                "description": "缓存回执为true；新执行为false。",
                "schema": {
                  "type": "boolean"
                }
              },
              "Location": {
                "description": "轮询或读取已创建资源；正文operationLocation/location相同。",
                "schema": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200,
                  "pattern": "^/api/v1/"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../contracts/public.schema.json#/$defs/ExportAccepted"
                },
                "example": {
                  "requestId": "11111111-1111-4111-8111-111111111111",
                  "export": {
                    "exportId": "11111111-1111-4111-8111-111111111111",
                    "sessionId": "11111111-1111-4111-8111-111111111111",
                    "sealedHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                    "status": "queued",
                    "createdAt": "2026-09-16T12:00:00Z",
                    "completedAt": null,
                    "artifact": null,
                    "errorCode": null
                  },
                  "operationLocation": "/api/v1/sessions/11111111-1111-4111-8111-111111111111/exports/11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-authorization": "local-launch-bound-commander",
        "x-idempotency": "required; unique key per launch/session; hash canonical body + method + path + runEpoch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../contracts/public.schema.json#/$defs/ExportRequest"
              },
              "example": {
                "sealedHash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                "format": "last-mile-review-json",
                "includePlayerStatements": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "LaunchBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque-local-launch-token",
        "description": "仅本机启动入口提供；绑定launch与commander能力。不得放URL、SSE query或日志。"
      }
    },
    "parameters": {
...
      "sessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "description": "",
        "schema": {
          "type": "string",
          "minLength": 36,
          "maxLength": 36,
          "format": "uuid"
        }
      },
...
    },
...
    "responses": {
      "Error400": {
        "description": "格式、参数或组合非法",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem400"
            },
            "example": {
              "type": "urn:last-mile:problem:invalid-request",
              "title": "格式、参数或组合非法",
              "status": 400,
              "code": "INVALID_REQUEST",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": false,
              "currentStateVersion": null,
              "detail": "格式、参数或组合非法",
              "violations": []
            }
          }
        }
      },
      "Error401": {
        "description": "本机启动令牌缺失或无效",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem401"
            },
            "example": {
              "type": "urn:last-mile:problem:unauthorized",
              "title": "本机启动令牌缺失或无效",
              "status": 401,
              "code": "UNAUTHORIZED",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": false,
              "currentStateVersion": null,
              "detail": "本机启动令牌缺失或无效",
              "violations": []
            }
          }
        }
      },
      "Error403": {
        "description": "授权会话、Origin或能力不允许",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem403"
            },
            "example": {
              "type": "urn:last-mile:problem:capability-denied",
              "title": "授权会话、Origin或能力不允许",
              "status": 403,
              "code": "CAPABILITY_DENIED",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": false,
              "currentStateVersion": null,
              "detail": "授权会话、Origin或能力不允许",
              "violations": []
            }
          }
        }
      },
      "Error404": {
        "description": "指定公开资源不存在或不属于会话",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem404"
            },
            "example": {
              "type": "urn:last-mile:problem:resource-not-found",
              "title": "指定公开资源不存在或不属于会话",
              "status": 404,
              "code": "RESOURCE_NOT_FOUND",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": false,
              "currentStateVersion": null,
              "detail": "指定公开资源不存在或不属于会话",
              "violations": []
            }
          }
        }
      },
      "Error409": {
        "description": "版本、场景、epoch或幂等键冲突",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem409"
            },
            "example": {
              "type": "urn:last-mile:problem:state-version-conflict",
              "title": "版本、场景、epoch或幂等键冲突",
              "status": 409,
              "code": "STATE_VERSION_CONFLICT",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": false,
              "currentStateVersion": null,
              "detail": "版本、场景、epoch或幂等键冲突",
              "violations": []
            }
          }
        }
      },
...
      "Error503": {
        "description": "服务或持久化暂不可用",
        "headers": {
          "X-Request-Id": {
            "description": "请求关联ID；所有响应均返回。",
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "format": "uuid"
            }
          },
          "Retry-After": {
            "description": "建议等待秒数；同逻辑请求保留Idempotency-Key。",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../contracts/public.schema.json#/$defs/Problem503"
            },
            "example": {
              "type": "urn:last-mile:problem:service-unavailable",
              "title": "服务或持久化暂不可用",
              "status": 503,
              "code": "SERVICE_UNAVAILABLE",
              "requestId": "11111111-1111-4111-8111-111111111111",
              "retryable": true,
              "currentStateVersion": null,
              "detail": "服务或持久化暂不可用",
              "violations": []
            }
          }
        }
      }
    }
  }
}